+++ src/ext/ed25519/ref10/sign.c
@@ -5,8 +5,8 @@
 #include "sc.h"
 
 int crypto_sign(
-  unsigned char *sm,unsigned long long *smlen,
-  const unsigned char *m,unsigned long long mlen,
+  unsigned char *sm,uint64_t *smlen,
+  const unsigned char *m,uint64_t mlen,
   const unsigned char *sk
 )
 {
